\name{get.best}
\alias{get.best}
\title{Get QAICc-best \code{fit.catch} object in a \code{fit.catch.compare} object}
\usage{
get.best(object)
}
\arguments{
  \item{object}{A \code{fit.catch.compare} object}
}
\value{
A \code{fit.catch} object
}
\description{
Get QAICc-best \code{fit.catch} object in a
\code{fit.catch.compare} object
}

