\docType{package}
\name{omnr.gillnet-package}
\alias{omnr.gillnet}
\alias{omnr.gillnet-package}
\title{Gillnet size selectivity analysis}
\description{
Gillnet size selectivity analyses for Ontario Ministry of
Natural Resources.
}
\details{
\tabular{ll}{ Package: \tab omnr.gillnet\cr Type: \tab
Package\cr Version: \tab 1.1\cr Date: \tab 2011-11-25\cr
License: \tab GPL-2\cr LazyLoad: \tab yes\cr LazyData: \tab
yes\cr } The most important functions are
\code{\link{make.catch}} and \code{\link{fit.catch}}.
}
\author{
Steve Walker

Maintainer: Steve Walker <steve.walker@utoronto.ca>
}
\references{
J.F. Bromaghin (2005) A versatile net selectivity model,
with application to Pacific salmon and freshwater species
of the Yukon River, Alaska. Fisheries Research 74: 157-168.

R.B. Millar & R.J. Fryer (1999) Estimating the
size-selection curves of towed gears, traps, nets and
hooks. Reviews in Fish Biology and Fisheries 9: 89-116.
}
\keyword{package}

