% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omnr.gillnet-package.R
\docType{package}
\name{omnr.gillnet-package}
\alias{omnr.gillnet-package}
\alias{omnr.gillnet}
\title{Gillnet size selectivity analysis}
\description{
Gillnet size selectivity analyses for Ontario Ministry of Natural Resources.
}
\details{
\tabular{ll}{ Package: \tab omnr.gillnet\cr Type: \tab Package\cr Version:
\tab 1.2\cr Date: \tab 2018-02-27\cr License: \tab GPL-2\cr LazyLoad: \tab
yes\cr LazyData: \tab yes\cr } The most important functions are
\code{\link{make.catch}} and \code{\link{fit.catch}}.
}
\references{
J.F. Bromaghin (2005) A versatile net selectivity model, with
application to Pacific salmon and freshwater species of the Yukon River,
Alaska. Fisheries Research 74: 157-168.

R.B. Millar & R.J. Fryer (1999) Estimating the size-selection curves of towed
gears, traps, nets and hooks. Reviews in Fish Biology and Fisheries 9:
89-116.
}
\author{
Steve Walker

Maintainer: Steve Walker <steve.walker@utoronto.ca>
}
\keyword{package}
